# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Jewish prayer times calculation and publisher management",
        "title": "Zmanim Lab API",
        "termsOfService": "https://zmanim-lab.com/terms",
        "contact": {
            "name": "Zmanim Lab Support",
            "email": "support@zmanim-lab.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/dsl/preview": {
            "post": {
                "description": "Calculates the result of a zmanim formula for a specific date and location, returning the time and calculation breakdown",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DSL"
                ],
                "summary": "Preview DSL formula",
                "parameters": [
                    {
                        "description": "Formula, date, and location",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.DSLPreviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calculation result with breakdown",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_handlers.DSLPreviewResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request or formula error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Location not found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dsl/preview-week": {
            "post": {
                "description": "Calculates a zmanim formula for 7 consecutive days starting from the specified date, including Hebrew dates and Shabbat/holiday markers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DSL"
                ],
                "summary": "Preview DSL formula for a week",
                "parameters": [
                    {
                        "description": "Formula, start date, and location",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.DSLPreviewWeekRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Weekly calculation results",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_handlers.DSLPreviewWeekResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Location not found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dsl/validate": {
            "post": {
                "description": "Validates a zmanim formula written in the DSL syntax, returning any errors and dependencies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DSL"
                ],
                "summary": "Validate DSL formula",
                "parameters": [
                    {
                        "description": "Formula to validate",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.DSLValidateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Validation result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_handlers.DSLValidateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_jcom-dev_zmanim-lab_internal_dsl.CalculationStep": {
            "type": "object",
            "properties": {
                "step": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "github_com_jcom-dev_zmanim-lab_internal_dsl.ValidationError": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "integer"
                },
                "line": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "suggestion": {
                    "type": "string"
                }
            }
        },
...
        "internal_handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/internal_handlers.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/internal_handlers.ResponseMeta"
                }
            }
        },
...
        "internal_handlers.DSLPreviewRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "ISO 8601 date (YYYY-MM-DD)",
                    "type": "string"
                },
                "elevation": {
                    "description": "Optional elevation in meters",
                    "type": "number"
                },
                "formula": {
                    "type": "string"
                },
                "latitude": {
                    "description": "Direct coordinates",
                    "type": "number"
                },
                "location_id": {
                    "description": "Optional: city/location ID",
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "timezone": {
                    "description": "e.g., \"America/New_York\"",
                    "type": "string"
                }
            }
        },
        "internal_handlers.DSLPreviewResponse": {
            "type": "object",
            "properties": {
                "breakdown": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_jcom-dev_zmanim-lab_internal_dsl.CalculationStep"
                    }
                },
                "result": {
                    "description": "Formatted time (HH:MM:SS)",
                    "type": "string"
                },
                "timestamp": {
                    "description": "Unix timestamp",
                    "type": "integer"
                }
            }
        },
        "internal_handlers.DSLPreviewWeekRequest": {
            "type": "object",
            "properties": {
                "elevation": {
                    "type": "number"
                },
                "formula": {
                    "type": "string"
                },
                "latitude": {
                    "description": "Direct coordinates",
                    "type": "number"
                },
                "location_id": {
                    "description": "Optional: city/location ID",
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "start_date": {
                    "description": "ISO 8601 date (YYYY-MM-DD)",
                    "type": "string"
                },
                "timezone": {
                    "description": "e.g., \"America/New_York\"",
                    "type": "string"
                }
            }
        },
        "internal_handlers.DSLPreviewWeekResponse": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handlers.DayPreview"
                    }
                }
            }
        },
        "internal_handlers.DSLValidateRequest": {
            "type": "object",
            "properties": {
                "formula": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.DSLValidateResponse": {
            "type": "object",
            "properties": {
                "dependencies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_jcom-dev_zmanim-lab_internal_dsl.ValidationError"
                    }
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "internal_handlers.DayPreview": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "events": {
                    "description": "Jewish holidays, Shabbat, etc.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "hebrew_date": {
                    "description": "Hebrew date string",
                    "type": "string"
                },
                "is_shabbat": {
                    "type": "boolean"
                },
                "is_yom_tov": {
                    "type": "boolean"
                },
                "result": {
                    "description": "Calculated time HH:MM:SS",
                    "type": "string"
                },
                "sunrise": {
                    "description": "HH:MM:SS",
                    "type": "string"
                },
                "sunset": {
                    "description": "HH:MM:SS",
                    "type": "string"
                }
            }
        },
...
        "internal_handlers.ResponseMeta": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token from Clerk authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}